Moving config_test.go to config/tests
Config_test.go now uses configurator as an executable to generate a proto locally.
Then, tests are run on the output of that local file to determine if configuration is correct for our repository.
Also includes a bugfix for configurator for reading to file
