+++ testgrid/cmd/configurator/BUILD.bazel
@@ -14,7 +14,6 @@ prow_image(name = "image")
 go_test(
     name = "go_default_test",
     srcs = [
-        "config_test.go",
         "main_test.go",
         "prow_test.go",
         "yaml2proto_test.go",
@@ -29,7 +28,6 @@ go_test(
         "//prow/apis/prowjobs/v1:go_default_library",
         "//prow/config:go_default_library",
         "//testgrid/config:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
     ],
 )
